requirements.txt
adafruit-circuitpython-neopixel==6.3.17
RPi.GPIO==0.7.1
rpi_ws281x==5.0.0
Flask==3.1.1